Mirclient: Avoid hardcoding the platform api choice
When built in by the debian scripts, a symbol PLATFORM_API_TOUCH would
be defined for relevant target platforms. Here in Qt, this does not apply.

On a correctly installed system, the UBUNTU_PLATFORM_API_BACKEND
variable is not required at runtime, since the system's value will be
read from a settings file under /etc.

Also, the previous hardcoding would mean that it could not be
overridden at runtime.

Change-Id: I24ddfaa254005b4113f3328b66edb1c6bbc509e2
Reviewed-by: Christian Stromme <christian.stromme@theqtcompany.com>