Windows/Direct2D: Handle screen changes after window creation.
Call QWindowSystemInterface::handleWindowScreenChanged() in
QWindowsIntegration::createPlatformWindow() should the effective
screen be different from the initial screen due to the geometry.

In order to avoid code duplication in the Direct2D plugin, fold the
helper method createWindowData() into
QWindowsIntegration::createPlatformWindow() and introduce a new
virtual helper to create the actual platform window for
QWindowsDirect2DIntegration to hook in.

Task-number: QTBUG-48379
Change-Id: I49c7a50a12245691414ab49bffe802239a87d2d6
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>