Cocoa: Forward rejected key events.
Forward rejected key events to the next responder
by checking the return value from QWindowSystemInterface
and calling the superclass event handler.

This is useful when Qt is running as a plugin in a
host application; the host can now react to key
events even if Qt has focus.

Qt will often not accept keyUp events, even if the
corresponding keyDown was accepted, for example in
the case of text controls. We don't want to forward
'bare' keyUps, so keep track of which keyDowns have
been seen.

Change-Id: I976448a5d305a657a0e91aeb271b158f8b598286
Task-number: QTBUG-45768
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@theqtcompany.com>