Add a file to the TESTDATA list
The file exists and it's listed in json.qrc (for Android) but it isn't
listed in TESTDATA.  This is only a problem when the target is a another
system as is the case with Qt for QNX.  One of the tests fails because
the file isn't deployed.

Noticed this while testing the changes for custom spacing of JSON
output.

Task-number: QTBUG-47437
Change-Id: I627592a5a225f50fdb8e3cdd6ac72dff43936ae5
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>